Improved testing of unit test results.
Added function ENSURE-INSTALL-RESULTS-SAME and used it in many of the
tests. In particular, replaced the test for success in
space-in-working-directory : test-1, allowing it to pass on Windows systems.

One thing I don't quite understand: When I made this change, I
discovered that installing moptilities also installs closer-mop in one
unit test, but not in another unit test. This wasn't tested for
before, so I don't know if it's expected or not. I adjusted the test
success conditions so that they both pass despite this anomaly.

darcs-hash:20071225202637-7fc34-cc17b8f3f241f94f174e8c1cc5953cfe02577f6d.gz