Removed binding of *DEFAULT-PATHNAME-DEFAULTS* in INSTALL-PACKAGE.
In LispWorks Personal 5.0.1, the call to DIRECTORY in INSTALL-PACKAGE
sometimes returns NIL when it shouldn't. The problem seems to related
to the binding and/or use of *DEFAULT-PATHNAME-DEFAULTS*. On
examination, it seems that binding this is really pointless anyway,
since it was also being passed explicitly to MAKE-PATHNAME. So I
replaced it with a lexical variable. That solved the problem that I
was encountering.

darcs-hash:20071227015725-7fc34-dfaa976cbc42471c3129e76aeec75dc3c404a14c.gz